add gz support for all formats
