+++ Analogy.LogViewer.Serilog/Parsers/CommonParser.cs
